Split runner into modules
