+++ core-runner/src/lib.rs
@@ -26,5 +26,7 @@ extern crate alloc;
 
 #[allow(dead_code)]
 mod builder;
+mod ext;
 pub mod runner;
+mod util;
 pub use runner::*;
