+++ core-runner/src/runner.rs
@@ -20,24 +20,24 @@
 
 use alloc::boxed::Box;
 use alloc::collections::{BTreeMap, BTreeSet};
-use alloc::string::String;
 use alloc::vec::Vec;
 use codec::{Decode, Encode};
 
-use gear_core::storage::WaitList;
 use gear_core::{
     env::Ext as EnvExt,
-    gas::{self, ChargeResult, GasCounter, GasCounterLimited},
+    gas::{self, GasCounter, GasCounterLimited},
     memory::{Memory, MemoryContext, PageNumber},
     message::{
         ExitCode, IncomingMessage, Message, MessageContext, MessageId, MessageIdGenerator,
-        OutgoingMessage, OutgoingPacket, ReplyMessage, ReplyPacket,
+        OutgoingMessage, ReplyMessage,
     },
     program::{Program, ProgramId},
-    storage::{MessageQueue, ProgramStorage, Storage},
+    storage::{MessageQueue, ProgramStorage, Storage, WaitList},
 };
-
 use gear_core_backend::Environment;
+
+use crate::ext::Ext;
+use crate::util::BlakeMessageIdGenerator;
 
 /// Runner configuration.
 #[derive(Clone, Debug, Decode, Encode)]
@@ -83,14 +83,6 @@ pub struct RunNextResult {
 }
 
 impl RunNextResult {
-    /// Result that notes that some log message had been handled, otherwise empty.
-    pub(crate) fn log() -> Self {
-        RunNextResult {
-            handled: 1,
-            ..Default::default()
-        }
-    }
-
     /// Result that notes that some failed program has been tried to run but nothing really happened.
     pub(crate) fn trap() -> Self {
         RunNextResult {
@@ -153,27 +145,6 @@ impl RunNextResult {
             }
         }
         false
-    }
-}
-
-/// Blake2 Message Id Generator
-pub struct BlakeMessageIdGenerator {
-    program_id: ProgramId,
-    nonce: u64,
-}
-
-impl gear_core::message::MessageIdGenerator for BlakeMessageIdGenerator {
-    fn next(&mut self) -> MessageId {
-        let mut data = self.program_id.as_slice().to_vec();
-        data.extend(&self.nonce.to_le_bytes());
-
-        self.nonce += 1;
-
-        MessageId::from_slice(blake2_rfc::blake2b::blake2b(32, &[], &data).as_bytes())
-    }
-
-    fn current(&self) -> u64 {
-        self.nonce
     }
 }
 
@@ -311,115 +282,101 @@ impl<MQ: MessageQueue, PS: ProgramStorage, WL: WaitList> Runner<MQ, PS, WL> {
             }
         };
 
-        if next_message.dest() == 0.into() {
-            match String::from_utf8(next_message.payload().to_vec()) {
-                Ok(s) => log::debug!("UTF-8 msg to /0: {}", s),
-                Err(_) => {
-                    log::debug!("msg to /0: {:?}", next_message.payload())
-                }
+        let gas_limit = next_message.gas_limit();
+        let next_message_source = next_message.source();
+        let next_message_dest = next_message.dest();
+
+        let mut program = match self.program_storage.get(next_message_dest) {
+            Some(program) => program,
+            None => {
+                // Reserve the entire `gas_limit` so that it is transferred to the addressee eventually
+                return RunNextResult::refund((next_message_source, next_message_dest, gas_limit));
             }
-            RunNextResult::log()
-        } else {
-            let gas_limit = next_message.gas_limit();
-            let next_message_source = next_message.source();
-            let next_message_dest = next_message.dest();
+        };
 
-            let mut program = match self.program_storage.get(next_message_dest) {
-                Some(program) => program,
-                None => {
-                    // Reserve the entire `gas_limit` so that it is transferred to the addressee eventually
-                    return RunNextResult::refund((
-                        next_message_source,
-                        next_message_dest,
-                        gas_limit,
-                    ));
-                }
-            };
-
-            if gas_limit > max_gas_limit {
-                // Re-queue the message to be processed in one of the following blocks
-                log::info!(
-                    "Message gas limit of {} exceeds the remaining block gas allowance of {}",
-                    gas_limit,
-                    max_gas_limit
-                );
-                self.message_queue.queue(next_message);
-                return RunNextResult::empty();
-            }
-
-            let instrumeted_code = match gas::instrument(program.code()) {
-                Ok(code) => code,
-                Err(err) => {
-                    log::debug!("Instrumentation error: {:?}", err);
-                    return RunNextResult::trap();
-                }
-            };
-
-            let allocations: BTreeSet<PageNumber> = program
-                .get_pages()
-                .iter()
-                .map(|(page_num, _)| *page_num)
-                .collect();
-
-            let mut context = self.create_context(allocations);
-            let next_message_id = next_message.id();
-
-            let mut run_result = run(
-                &mut self.env,
-                &mut context,
-                &instrumeted_code,
-                &mut program,
-                if next_message.reply().is_some() {
-                    EntryPoint::HandleReply
-                } else {
-                    EntryPoint::Handle
-                },
-                &next_message.into(),
+        if gas_limit > max_gas_limit {
+            // Re-queue the message to be processed in one of the following blocks
+            log::info!(
+                "Message gas limit of {} exceeds the remaining block gas allowance of {}",
                 gas_limit,
+                max_gas_limit
             );
-
-            if run_result.outcome.was_trap() {
-                // In case of trap, we generate trap reply message
-                let program_id = program.id();
-                let nonce = program.fetch_inc_message_nonce();
-                let trap_message_id = self.next_message_id(program_id, nonce);
-
-                self.message_queue.queue(Message {
-                    id: trap_message_id,
-                    source: program_id,
-                    dest: next_message_source,
-                    payload: vec![].into(),
-                    gas_limit: run_result.gas_left,
-                    value: 0,
-                    reply: Some((next_message_id, EXIT_CODE_PANIC)),
-                });
-            }
-
-            if let Some(waiting_msg) = run_result.waiting.take() {
-                self.wait_list.insert(waiting_msg.id, waiting_msg);
-            }
-
-            if let Some((gas, waker_id)) = run_result.awakening {
-                if let Some(mut msg) = self.wait_list.remove(waker_id) {
-                    // Increase gas available to the message
-                    msg.gas_limit += gas;
-                    context.message_buf.push(msg);
-                }
-            }
-
-            let result = RunNextResult::from_single(
-                next_message_id,
-                next_message_source,
-                next_message_dest,
-                run_result,
-            );
-
-            self.message_queue
-                .queue_many(context.message_buf.drain(..).collect());
-            self.program_storage.set(program);
-
-            result
+            self.message_queue.queue(next_message);
+            return RunNextResult::empty();
         }
+
+        let instrumeted_code = match gas::instrument(program.code()) {
+            Ok(code) => code,
+            Err(err) => {
+                log::debug!("Instrumentation error: {:?}", err);
+                return RunNextResult::trap();
+            }
+        };
+
+        let allocations: BTreeSet<PageNumber> = program
+            .get_pages()
+            .iter()
+            .map(|(page_num, _)| *page_num)
+            .collect();
+
+        let mut context = self.create_context(allocations);
+        let next_message_id = next_message.id();
+
+        let mut run_result = run(
+            &mut self.env,
+            &mut context,
+            &instrumeted_code,
+            &mut program,
+            if next_message.reply().is_some() {
+                EntryPoint::HandleReply
+            } else {
+                EntryPoint::Handle
+            },
+            &next_message.into(),
+            gas_limit,
+        );
+
+        if run_result.outcome.was_trap() {
+            // In case of trap, we generate trap reply message
+            let program_id = program.id();
+            let nonce = program.fetch_inc_message_nonce();
+            let trap_message_id = self.next_message_id(program_id, nonce);
+
+            self.message_queue.queue(Message {
+                id: trap_message_id,
+                source: program_id,
+                dest: next_message_source,
+                payload: vec![].into(),
+                gas_limit: run_result.gas_left,
+                value: 0,
+                reply: Some((next_message_id, EXIT_CODE_PANIC)),
+            });
+        }
+
+        if let Some(waiting_msg) = run_result.waiting.take() {
+            self.wait_list.insert(waiting_msg.id, waiting_msg);
+        }
+
+        if let Some((gas, waker_id)) = run_result.awakening {
+            if let Some(mut msg) = self.wait_list.remove(waker_id) {
+                // Increase gas available to the message
+                msg.gas_limit += gas;
+                context.message_buf.push(msg);
+            }
+        }
+
+        let result = RunNextResult::from_single(
+            next_message_id,
+            next_message_source,
+            next_message_dest,
+            run_result,
+        );
+
+        self.message_queue
+            .queue_many(context.message_buf.drain(..).collect());
+        self.program_storage.set(program);
+
+        result
     }
 
     /// Drop this runner.
@@ -641,192 +598,6 @@ pub struct RunResult {
     pub gas_requested: u64,
     /// Run outcome (trap/succes).
     pub outcome: ExecutionOutcome,
-}
-
-struct Ext {
-    memory_context: MemoryContext,
-    messages: MessageContext<BlakeMessageIdGenerator>,
-    gas_counter: Box<dyn GasCounter>,
-    gas_requested: u64,
-    alloc_cost: u64,
-    last_error_returned: Option<&'static str>,
-}
-
-impl Ext {
-    fn return_with_tracing<T>(
-        &mut self,
-        result: Result<T, &'static str>,
-    ) -> Result<T, &'static str> {
-        match result {
-            Ok(result) => Ok(result),
-            Err(error_string) => {
-                self.last_error_returned = Some(error_string);
-                Err(error_string)
-            }
-        }
-    }
-}
-
-impl EnvExt for Ext {
-    fn alloc(&mut self, pages: PageNumber) -> Result<PageNumber, &'static str> {
-        self.gas(pages.raw() * self.alloc_cost as u32)?;
-        let result = self
-            .memory_context
-            .alloc(pages)
-            .map_err(|_e| "Allocation error");
-
-        self.return_with_tracing(result)
-    }
-
-    fn send(&mut self, msg: OutgoingPacket) -> Result<MessageId, &'static str> {
-        if self.gas_counter.charge(msg.gas_limit()) != ChargeResult::Enough {
-            return Err("Gas limit exceeded while trying to send message");
-        }
-        let result = self.messages.send(msg).map_err(|_e| "Message send error");
-
-        self.return_with_tracing(result)
-    }
-
-    fn send_init(&mut self, msg: OutgoingPacket) -> Result<usize, &'static str> {
-        let result = self
-            .messages
-            .send_init(msg)
-            .map_err(|_e| "Message init error");
-
-        self.return_with_tracing(result)
-    }
-
-    fn send_push(&mut self, handle: usize, buffer: &[u8]) -> Result<(), &'static str> {
-        let result = self
-            .messages
-            .send_push(handle, buffer)
-            .map_err(|_e| "Payload push error");
-
-        self.return_with_tracing(result)
-    }
-
-    fn reply_push(&mut self, buffer: &[u8]) -> Result<(), &'static str> {
-        let result = self
-            .messages
-            .reply_push(buffer)
-            .map_err(|_e| "Reply payload push error");
-
-        self.return_with_tracing(result)
-    }
-
-    fn send_commit(&mut self, handle: usize) -> Result<MessageId, &'static str> {
-        {
-            let gas_limit = match self
-                .messages
-                .get_gas_limit(handle)
-                .map_err(|_e| "Message commit error")
-            {
-                Ok(gas) => gas,
-                Err(_e) => {
-                    return self.return_with_tracing(Err("No message to commit"));
-                }
-            };
-
-            if self.gas_counter.charge(gas_limit) != ChargeResult::Enough {
-                return self
-                    .return_with_tracing(Err("Gas limit exceeded while trying to send message"));
-            }
-        }
-
-        let result = self
-            .messages
-            .send_commit(handle)
-            .map_err(|_e| "Message commit error");
-
-        self.return_with_tracing(result)
-    }
-
-    fn reply(&mut self, msg: ReplyPacket) -> Result<(), &'static str> {
-        let result = self.messages.reply(msg).map_err(|_e| "Reply error");
-
-        self.return_with_tracing(result)
-    }
-
-    fn reply_to(&self) -> Option<(MessageId, ExitCode)> {
-        self.messages.current().reply()
-    }
-
-    fn source(&mut self) -> ProgramId {
-        self.messages.current().source()
-    }
-
-    fn message_id(&mut self) -> MessageId {
-        self.messages.current().id()
-    }
-
-    fn free(&mut self, ptr: PageNumber) -> Result<(), &'static str> {
-        let result = self.memory_context.free(ptr).map_err(|_e| "Free error");
-        self.return_with_tracing(result)
-    }
-
-    fn debug(&mut self, data: &str) -> Result<(), &'static str> {
-        log::debug!("DEBUG: {}", data);
-        Ok(())
-    }
-
-    fn set_mem(&mut self, ptr: usize, val: &[u8]) {
-        self.memory_context
-            .memory()
-            .write(ptr, val)
-            // TODO: remove and propagate error, issue #97
-            .expect("Memory out of bounds.");
-    }
-
-    fn get_mem(&self, ptr: usize, buffer: &mut [u8]) {
-        self.memory_context.memory().read(ptr, buffer);
-    }
-
-    fn msg(&mut self) -> &[u8] {
-        self.messages.current().payload()
-    }
-
-    fn gas(&mut self, val: u32) -> Result<(), &'static str> {
-        if self.gas_counter.charge(val as u64) == ChargeResult::Enough {
-            Ok(())
-        } else {
-            self.return_with_tracing(Err("Gas limit exceeded"))
-        }
-    }
-
-    fn gas_available(&mut self) -> u64 {
-        self.gas_counter.left()
-    }
-
-    fn charge(&mut self, gas: u64) -> Result<(), &'static str> {
-        if self.gas_counter.charge(gas) == ChargeResult::Enough {
-            self.gas_requested += gas;
-            Ok(())
-        } else {
-            self.return_with_tracing(Err("Gas limit exceeded"))
-        }
-    }
-
-    fn value(&self) -> u128 {
-        self.messages.current().value()
-    }
-
-    fn wait(&mut self) -> Result<(), &'static str> {
-        let result = self
-            .messages
-            .wait()
-            .map_err(|_| "Unable to add the message to the wait list");
-
-        self.return_with_tracing(result)
-    }
-
-    fn wake(&mut self, waker_id: MessageId) -> Result<(), &'static str> {
-        let result = self
-            .messages
-            .wake(waker_id)
-            .map_err(|_| "Unable to mark the message to be woken");
-
-        self.return_with_tracing(result)
-    }
 }
 
 fn run(
