+++ core/src/storage.rs
@@ -35,6 +35,9 @@ pub trait ProgramStorage: Default {
     /// Store program in the storage.
     fn set(&mut self, program: Program) -> Option<Program>;
 
+    /// Check if program exists.
+    fn exists(&self, id: ProgramId) -> bool;
+
     /// Remove the program from the storage.
     fn remove(&mut self, id: ProgramId) -> Option<Program>;
 }
@@ -59,6 +62,10 @@ impl ProgramStorage for InMemoryProgramStorage {
 
     fn set(&mut self, program: Program) -> Option<Program> {
         self.inner.insert(program.id(), program)
+    }
+
+    fn exists(&self, id: ProgramId) -> bool {
+        self.inner.contains_key(&id)
     }
 
     fn remove(&mut self, id: ProgramId) -> Option<Program> {
@@ -311,6 +318,38 @@ mod tests {
     /// Test that InMemoryMessageQueue works correctly
     fn message_queue_interaction() {
         use crate::message::Payload;
+
+        let wat = r#"
+            (module
+                (import "env" "gr_reply_to"  (func $gr_reply_to (param i32)))
+                (import "env" "memory" (memory 2))
+                (export "handle" (func $handle))
+                (export "handle_reply" (func $handle))
+                (export "init" (func $init))
+                (func $handle
+                    i32.const 65536
+                    call $gr_reply_to
+                )
+                (func $handle_reply
+                    i32.const 65536
+                    call $gr_reply_to
+                )
+                (func $init)
+            )"#;
+
+        let binary: Vec<u8> = parse_wat(wat);
+
+        // Initialization of some ProgramIds
+        let id1 = ProgramId::from(1);
+
+        let id2 = ProgramId::from(2);
+
+        // Initialization of InMemoryProgramStorage with our custom vec<Program>
+        let program_storage: InMemoryProgramStorage = vec![
+            Program::new(id1, binary.clone(), Default::default()).expect("err create program"),
+            Program::new(id2, binary.clone(), Default::default()).expect("err create program"),
+        ]
+        .into();
 
         // Initialization of empty InMemoryMessageQueue
         let mut message_queue = InMemoryMessageQueue::new();
